[REF] typ_stock: Add flag to indicate if a purchase order line must be deleted. (#272)
* [REF] typ_stock: Add in context of action server a flag to indicate if a purchase order line must be deleted.

* [ADD] typ_stock: Add tests to validate appropriate operation of action server respect to purchase and procurement.

* [REF] typ_stock: Remove run action server.

+++ typ_stock/models/orderpoint_procurement.py
@@ -119,6 +119,8 @@ class PurchaseOrderLine(models.Model):
 
     @api.multi
     def unlink(self):
+        if self._context.get('not_delete_purchase_line', False):
+            return True
         proc_ids = self.mapped('procurement_ids')
         res = super(PurchaseOrderLine, self).unlink()
         if proc_ids:
